requirements.txt
asgiref==3.11.0
Django==6.0
mysqlclient==2.2.7
sqlparse==0.5.5
tzdata==2025.3
PyJWT==2.10.1
xhtml2pdf
django-storages
boto3
